requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-datatables==0.6.0
Pygments==2.7.1
pytz==2020.1
sqlparse==0.3.1
